# src/infra/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação Swagger do Projeto Liven",
        "version": "1.0.0",
        "title": "Projeto Liven",
        "contact": {
            "email": "keltonmof@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/createUser": {
            "post": {
                "summary": "Cadastra um usuário",
                "description": "Cadastra um usuário com base nos parâmetros enviados",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Usuário",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "User": {
                                    "type": "object",
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário incluído",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/createUserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Login do usuário",
                "description": "Com base no login e senha retorna um token de acesso",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Login",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "User": {
                                    "type": "object",
                                    "$ref": "#/definitions/login"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login efetuado com sucesso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/updateUser": {
            "post": {
                "summary": "Atualiza usuário logado",
                "description": "Atualiza o usuário logado com base nos parâmetros enviados",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Usuário",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "User": {
                                    "type": "object",
                                    "$ref": "#/definitions/User"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário atualizado",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/updateUserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/deleteUser": {
            "post": {
                "summary": "Deleta usuário logado",
                "description": "Deleta o usuário logado com base no token",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/deleteUserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/createAddress": {
            "post": {
                "summary": "Cria um endereço",
                "description": "Cria um endereço para o usuário logado com base no token",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Address",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Address": {
                                    "type": "object",
                                    "$ref": "#/definitions/Address"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço criado",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/createAddress"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/updateAddress": {
            "post": {
                "summary": "Atualiza um endereço",
                "description": "Atualiza um endereço caso o usuário tenha permissão sobre ele",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Address",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Address": {
                                    "type": "object",
                                    "$ref": "#/definitions/updateAddress"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço atualizado com sucesso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/updateUserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/deleteAddress": {
            "post": {
                "summary": "Atualiza um endereço",
                "description": "Atualiza um endereço caso o usuário tenha permissão sobre ele",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Address apenas com código",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Address": {
                                    "type": "object",
                                    "$ref": "#/definitions/deleteAddress"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço excluído com sucesso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/deleteUserResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/getAllInfo": {
            "get": {
                "summary": "Lista todas as informações",
                "description": "Lista todas as informações de um usuário específico e seus endereços vinculados",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados recuperados com suceso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "createUserResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "job": {
                    "type": "string"
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "login": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "loginResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "updateUserResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "updated": {
                    "type": "boolean"
                }
            }
        },
        "deleteUser": {
            "type": "object",
            "properties": {}
        },
        "deleteUserResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "deleted": {
                    "type": "boolean"
                }
            }
        },
        "Address": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "zipCode": {
                    "type": "string"
                },
                "aditionalInformation": {
                    "type": "string"
                }
            }
        },
        "createAddress": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "addressCode": {
                    "type": "integer"
                }
            }
        },
        "updateAddress": {
            "type": "object",
            "properties": {
                "addressCode": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "zipCode": {
                    "type": "string"
                },
                "aditionalInformation": {
                    "type": "string"
                }
            }
        },
        "deleteAddress": {
            "type": "object",
            "properties": {
                "addressCode": {
                    "type": "integer"
                }
            }
        }
    }
}